// src/types/bicycle.types.ts
export type TProduct = {
  brand: string;
  category: string;
  createdAt: string;
  image: string;
  model: string;
  name: string;
  price: number;
  productDetails: string;
  stock: number;
  updatedAt: string;
  __v: number;
  _id: string;
};

export type TQueryParams = {
  name: string;
  value: string;
};